requirements.txt
setuptools>=68.0.0
pyspark>=3.4.0
delta-spark>=2.4.0
